Fix secondary title render issue with markdown-it.
Previously the cookie policy and privacy notice page doesn't render first title appeared in the .md file properly. This might be related to how markdown-it renders a .md file internally. Because spaces are quite important for markdown, I add a newline at the top of these two md file and it now works.